aula-18/livraria/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
Django==4.2.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-yasg==1.21.6
flasgger==0.9.7.1
Flask==2.3.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mistune==2.0.5
model-bakery==1.12.0
packaging==23.1
psycopg==3.1.9
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.3
Werkzeug==2.3.4
